Adding our custom made buttons to the menu

+++ PlatformerTutorial/src/entities/Entity.java
@@ -25,7 +25,7 @@ public abstract class Entity {
 
 	}
 
-	protected void initHitbox(float x, float y, float width, float height) {
+	protected void initHitbox(float x, float y, int width, int height) {
 		hitbox = new Rectangle2D.Float(x, y, width, height);
 	}
